backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pyjwt==2.8.0
python-multipart==0.0.9
tensorflow==2.16.1
numpy==1.26.4
scikit-learn==1.5.0
pandas==2.2.2
